+++ src/main/java/com/github/hcsp/inheritance/People.java
@@ -1,0 +1,14 @@
+package com.github.hcsp.inheritance;
+
+public class People extends Animal {
+    public int age;
+
+    People(String name, int age) {
+        super(name);
+        this.age = age;
+    }
+
+    public void sayMyAge() {
+        System.out.println("我的年龄是" + this.age);
+    }
+}
